requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==3.2.2; python_version >= '3.6'
beautifulsoup4==4.11.1; python_version >= '3.6'
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3; python_version >= '3.7'
colorama==0.4.4; platform_system == 'Windows'
cryptography==37.0.2
flask-bcrypt==1.0.1
flask==2.1.2
idna==3.3; python_version >= '3'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
pycparser==2.21
pymysql==1.0.2
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.1.2; python_version >= '3.7'
xlsxwriter==3.0.3
